pyproject.toml
[project]
name = "quickapply"
version = "0.1.0"
description = "Local JobSpy-powered job monitor with tracker and Codex tailoring workflow"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "bleach>=6.3.0",
    "fastapi>=0.135.3",
    "flask>=3.1.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "markdown>=3.10.2",
    "python-dateutil>=2.9.0.post0",
    "python-jobspy>=1.1.82",
    "python-multipart>=0.0.26",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.5",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
